Lanuk-Backend/pyproject.toml
[project]
name = "Lanuk-Backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Falk Stankat",email = "falk.stankat@disy.net"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "openai (>=1.61.1,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "reportlab (>=4.3.1,<5.0.0)",
    "cairosvg (>=2.7.1,<3.0.0)",
    "flask[async] (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "tavily-python (>=0.5.4,<0.6.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "langchain-openai (>=0.3.13,<0.4.0)",
    "langchain-fireworks (>=0.2.9,<0.3.0)",
    "langchain-deepseek (>=0.1.3,<0.2.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "langgraph (>=0.4.1,<0.5.0)",
    "langchain-anthropic (>=0.3.12,<0.4.0)",
    "pyairtable (>=3.1.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "scipy (>=1.16.0,<2.0.0)",
    "odfpy (>=1.4.1,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
